Move clickingOnSkipButtonOpensLoginScreen() test
