+++ app/src/androidTest/java/com/github/geohunt/app/ui/components/tutorial/WelcomeScreenTest.kt
@@ -8,8 +8,11 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.test.core.app.launchActivity
 import androidx.test.espresso.intent.Intents
+import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.github.geohunt.app.LoginActivity
 import com.github.geohunt.app.TutorialActivity
+import org.hamcrest.Matchers
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
@@ -21,12 +24,12 @@ class WelcomeScreenTest {
     @get:Rule
     val composeTestRule = createAndroidComposeRule<TutorialActivity>()
 
-    val PREFERENCES_FILE = "Preferences"
+    private val PREFERENCES_FILE = "preferences"
 
     @Before
     fun setup(){
         val settings: SharedPreferences = composeTestRule.activity.getSharedPreferences(PREFERENCES_FILE, 0)
-        settings.edit().putBoolean("first_application_open", true).commit()
+        settings.edit().putBoolean("first_application_open", true).apply()
     }
 
     @After
@@ -53,6 +56,24 @@ class WelcomeScreenTest {
             .assertExists()
             .assertHasClickAction()
 
+        Intents.release()
+    }
+
+    @Test
+    fun clickingOnSkipButtonOpensLoginScreen() {
+        Intents.init()
+        launchActivity<TutorialActivity>()
+
+        composeTestRule
+            .onNodeWithText("GET STARTED")
+            .performClick()
+
+        composeTestRule
+            .onNodeWithTag("Skip button")
+            .assertHasClickAction()
+            .performClick()
+
+        Intents.intended(Matchers.allOf(IntentMatchers.hasComponent(LoginActivity::class.java.name)))
         Intents.release()
     }
